move Page.js to Page/index.jsx
