+++ src/index.js
@@ -4,7 +4,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import * as serviceWorker from "./serviceWorker";
 
-import App from "./App";
+import App from "./components/App";
 import FirebaseContext from "./config/Context";
 import Firebase from "./config/Firebase";
